Moved FAQ and refactored path into constant
