+++ src/main.tsx
@@ -6,7 +6,7 @@ import HomeView from './components/HomeView/HomeView'
 import PricingView from './components/PricingView/PricingView'
 import AboutView from './components/AboutView/AboutView'
 import DocsView from './components/DocsView/DocsView'
-import FAQ from './components/DocsView/FAQ/FAQ'
+import FAQ from './components/DocsView/components/FAQ/FAQ'
 import * as FastClick from 'fastclick'
 import QuickstartPage from './components/DocsView/pages/quickstart/QuickstartPage'
 import ReferencePage from './components/DocsView/pages/reference/ReferencePage'
